requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dnspython==2.8.0
fastapi==0.128.0
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
motor==3.7.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
python-multipart==0.0.22
setuptools==70.2.0
starlette==0.50.0
sympy==1.14.0
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
